+++ examples/with-supertokens/pages/_app.js
@@ -1,0 +1,41 @@
+import '../styles/globals.css'
+import React from 'react'
+import { useEffect } from 'react'
+import SuperTokensReact from 'supertokens-auth-react'
+import * as SuperTokensConfig from '../config/frontendConfig'
+import Session from 'supertokens-auth-react/recipe/session'
+import { redirectToAuth } from 'supertokens-auth-react/recipe/thirdpartyemailpassword'
+
+async function initNode() {
+  const supertokensNode = await import('supertokens-node')
+  const { backendConfig } = await import('../config/backendConfig')
+  supertokensNode.init(backendConfig())
+}
+
+if (typeof window !== 'undefined') {
+  SuperTokensReact.init(SuperTokensConfig.frontendConfig())
+} else {
+  initNode().catch(console.error)
+}
+
+function MyApp({ Component, pageProps }) {
+  useEffect(() => {
+    async function doRefresh() {
+      if (pageProps.fromSupertokens === 'needs-refresh') {
+        if (await Session.attemptRefreshingSession()) {
+          location.reload()
+        } else {
+          // user has been logged out
+          redirectToAuth()
+        }
+      }
+    }
+    doRefresh()
+  }, [pageProps.fromSupertokens])
+  if (pageProps.fromSupertokens === 'needs-refresh') {
+    return null
+  }
+  return <Component {...pageProps} />
+}
+
+export default MyApp
