+++ examples/with-supertokens/pages/api/auth/[[...path]].js
@@ -2,9 +2,9 @@
 import { superTokensNextWrapper } from 'supertokens-node/nextjs'
 import supertokens from 'supertokens-node'
 import { middleware } from 'supertokens-node/framework/express'
-import * as SuperTokensConfig from '../../../config/supertokensConfig'
+import { backendConfig } from '../../../config/backendConfig'
 
-supertokens.init(SuperTokensConfig.backendConfig())
+supertokens.init(backendConfig())
 
 export default async function superTokens(req, res) {
   await superTokensNextWrapper(
